+++ hata/ext/rpc/__init__.py
@@ -4,12 +4,14 @@ from .client import *
 from .command_handling import *
 from .constants import *
 from .dispatch_handling import *
+from .event_types import *
 from .exceptions import *
 from .preinstanced import *
 from .rich_voice_state import *
 from .user_voice_settings import *
 from .utils import *
 from .voice_settings import *
+from .voice_connection_status import *
 
 __all__ = (
     *authenticate.__all__,
@@ -18,10 +20,12 @@ __all__ = (
     *command_handling.__all__,
     *constants.__all__,
     *dispatch_handling.__all__,
+    *event_types.__all__,
     *exceptions.__all__,
     *preinstanced.__all__,
     *rich_voice_state.__all__,
     *user_voice_settings.__all__,
     *utils.__all__,
     *voice_settings.__all__,
+    *voice_connection_status.__all__,
 )