+++ hata/ext/rpc/rich_voice_state.py
@@ -6,7 +6,7 @@ from .user_voice_settings import AudioBalance
 
 class RichVoiceState:
     """
-    Rich voice state containing user, it's voice state and other related attributes.
+    Rich voice state containing user, and it's voice state and other related attributes.
     
     Attributes
     ----------